Moved macwire and guice in the same project.
Ref: SCALA-93
