--- play-scala/compile-time-dependency-injection/README.md
@@ -1,6 +1,0 @@
-### Dependency Injection with Play
-Sample code to demonstrate using Play with MacWire compile-time dependency injection.
-
-#### Build/Run
-From sbt, simply execute the command "run". By default, the web server will be
-started on port 9000.
