--- play-scala/compile-time-dependency-injection/app/models/Order.scala
@@ -1,9 +1,0 @@
-package models
-
-import play.api.libs.json.Json
-
-object Order {
-  implicit val writes = Json.writes[Order]
-}
-
-case class Order(id: Long, userId: Long, date: Long, isEnterprise: Boolean)
