--- play-scala/compile-time-dependency-injection/app/models/User.scala
@@ -1,9 +1,0 @@
-package models
-
-import play.api.libs.json.Json
-
-case class User(id: Long, name: String)
-
-object User {
-  implicit val writes = Json.writes[User]
-}
