--- play-scala/compile-time-dependency-injection/app/services/UserService.scala
@@ -1,14 +1,0 @@
-package services
-
-import models.User
-
-class UserService {
-
-  def getAll(): Seq[User] = {
-    Seq(
-      User(1L, "John Doe"),
-      User(2L, "Sam Doe")
-    )
-  }
-
-}
