--- play-scala/compile-time-dependency-injection/conf/application.conf
@@ -1,1 +1,0 @@
-play.application.loader = "loader.AppApplicationLoader"
