--- play-scala/compile-time-dependency-injection/conf/routes
@@ -1,8 +1,0 @@
-# Routes
-# This file defines all application routes (Higher priority routes first)
-# https://www.playframework.com/documentation/latest/ScalaRouting
-# ~~~~
-
-# An example controller showing a sample home page
-GET     /users                           controllers.UserController.index()
-
