--- play-scala/compile-time-dependency-injection/project/plugins.sbt
@@ -1,3 +1,0 @@
-addSbtPlugin("com.typesafe.play" % "sbt-plugin" % "2.8.8")
-addSbtPlugin("org.foundweekends.giter8" % "sbt-giter8-scaffold" % "0.11.0")
-addSbtPlugin("org.scalameta" % "sbt-scalafmt" % "2.4.2")
