--- play-scala/compile-time-dependency-injection/test/com/baeldung/components/ServiceComponents.scala
@@ -1,15 +1,0 @@
-package com.baeldung.components
-
-import com.softwaremill.macwire._
-
-class ServiceWithRemoteCall(remoteApi: RemoteApi) {
-  def call(): String = remoteApi.remoteCall()
-}
-
-trait ServiceComponents {
-
-  self: ApiComponents =>
-
-  lazy val serviceWithRemoteCall: ServiceWithRemoteCall = wire[ServiceWithRemoteCall]
-
-}
