+++ play-scala/dependency-injection/README.md
@@ -1,5 +1,8 @@
 ### Dependency Injection with Play
-Sample code to demonstrate using Play with Google Guice dependency injection.
+Sample code to demonstrate using Play with Google Guice or MacWire dependency injection.
+
+#### Select Dependency Injection Framework
+You can choose DI framework by commenting/uncommenting the relevant sections in application.conf and routes files.
 
 #### Build/Run
 From sbt, simply execute the command "run". By default, the web server will be
