+++ play-scala/dependency-injection/app/macwire/services/UserService.scala
@@ -1,0 +1,14 @@
+package macwire.services
+
+import models.User
+
+class UserService {
+
+  def getAll(): Seq[User] = {
+    Seq(
+      User(1L, "John Doe"),
+      User(2L, "Sam Doe")
+    )
+  }
+
+}
