--- play-scala/dependency-injection/app/services/OrderValidationService.scala
@@ -1,21 +1,0 @@
-package services
-
-import models.Order
-
-trait OrderValidationService {
-  def validate(order: Order): Boolean
-}
-
-class BusinessOrderValidationService extends OrderValidationService {
-  override def validate(order: Order): Boolean = {
-    println("Business order validation")
-    true
-  }
-}
-
-class EnterpriseOrderValidationService extends OrderValidationService {
-  override def validate(order: Order): Boolean = {
-    println("Enterprise order validation")
-    true
-  }
-}
