+++ play-scala/dependency-injection/build.sbt
@@ -1,5 +1,4 @@
 name := """play-dependency-injection"""
-organization := "com.baeldung"
 
 version := "1.0-SNAPSHOT"
 
@@ -9,4 +8,5 @@ scalaVersion := "2.13.6"
 
 libraryDependencies += guice
 libraryDependencies += "org.scalatestplus.play" %% "scalatestplus-play" % "5.0.0" % Test
-
+libraryDependencies += "com.softwaremill.macwire" %% "macros" % "2.4.0" % Provided
+libraryDependencies += "com.softwaremill.macwire" %% "util" % "2.4.0"
