+++ play-scala/dependency-injection/conf/application.conf
@@ -1,4 +1,8 @@
-play {
-  modules.enabled += "modules.UserModule"
-  modules.enabled += "modules.OrderModule"
-}
+# Guice
+# play {
+#   guice.modules.enabled += "guice.modules.UserModule"
+#  guice.modules.enabled += "guice.modules.OrderModule"
+# }
+
+# MacWire
+play.application.loader = "macwire.loader.AppApplicationLoader"
