+++ play-scala/dependency-injection/test/macwire/components/ServiceComponents.scala
@@ -1,0 +1,15 @@
+package macwire.components
+
+import com.softwaremill.macwire._
+
+class ServiceWithRemoteCall(remoteApi: RemoteApi) {
+  def call(): String = remoteApi.remoteCall()
+}
+
+trait ServiceComponents {
+
+  self: ApiComponents =>
+
+  lazy val serviceWithRemoteCall: ServiceWithRemoteCall = wire[ServiceWithRemoteCall]
+
+}
